This seems to work flawlessly.
I will check again with 50 micro delay for latch.
Then attempt to decrease the 5 sec on time to 1 milli.
if that is unstable I will attempt pull up all data, clock and latches to high
using 4.7 k resistor (would love to test with 10k using oscilioscope but there is no time).

SOOOO weird that latch drain does not need a delay.
Maybe it is because output enable is disabled?