It is functioning. Made SW_SERR act as SW_WR and vice versa. Because it was annoying to display binary addition otherwise. Could have fit the SW_WR and SW_MEM 1-5mm higher and it would have been more erogonomical. Tomorrow I will solder a rev-1 without ground fill. Using 33Ohm instead of 51 ohm on datalines. Also using 220 instead of 330 ohm for the segments and LEDs. Also using blue instead of red LEDs.
Will not mount switches on this test, I will also use minimal amount of tin on the 33 ohm resistors, since I will probably test 51, and then 220 ohms if it is unstable.

All in all approaching demonstrateable version.
Could maybe restructure parts of code into seperate files or libraries.
But I will leave that to the date I add this repo to the monorepo.